data/requirements.txt
anthropic
pypdf
pillow
pdf2image
